waiv797::refactor object page feed component
