+++ src/client/object/ObjectFeed/ObjectFeed-nightmode.less
@@ -1,8 +1,7 @@
-@import (reference) '../styles/custom.less';
-@import (reference) '../styles/modules/variables.less';
+@import (reference) '../../styles/custom.less';
 
 .nightmode {
-  .object-profile {
+  .object-feed {
     &__row {
       background: @grey-nero !important;
       border-color: @normal-grey !important;
