+++ src/client/object/ObjectFeed/ObjectFeed.less
@@ -1,7 +1,6 @@
-@import (reference) '../styles/custom.less';
-@import (reference) '../styles/modules/variables.less';
+@import (reference) '../../styles/custom.less';
 
-.object-profile {
+.object-feed {
   &__row {
     display: flex;
     background: @white;
@@ -14,4 +13,4 @@
   }
 }
 
-@import 'ObjectProfile-nightmode';
+@import 'ObjectFeed-nightmode';
